Bug fixes and refactoring
Split UToolData into UToolData for the runtime module and UEditorData for the editor module. EditorData now contains all the data that is specific to the editors, rather than it being in the tool data.
Introduced SLightEditorWidget as a base for SLightControlTool and SDMXControlTool to remove large chunks of code duplication.
Fixed the saturation gradient not updating after a hue transaction. Fixed how light property transactions are reflected in the UI - there is no longer a delay after the transaction for the changes to be displayed.