Progress on tracking object
Changed MovementTargetComponentReference to TransformationTargetComponentReferene, along with relating variables and members.
Changed raw USceneComponent* to use TWeakObjectPtr to deal with possible garbage coillection. (Might need to become a TSharedObjectPtr if the movement target should never be deleted or garbage collected when referenced by the physicalobjecttracker references them.
Tested with a tracker for movement. Need to detect editing of component reference so it can be "extracted" as it uses a string to point to a component, but now the actual component either needs to be extracted each update or only when changed to have the tracker move the right object.