Add Tracker Serial ID as data asset.
Changed Tracker Serial ID variable to be a data asset in order to be able to reference the same asset in multiple scene and only associate the trackers by serial ID once for each shoot.
Originally trackers could be set on the PhysicalObjectTrackingComponent through a button. Now this button has been moved to the context menu of the PhysicalObjectTrackerSerialId asset.