Expand Debug visualization for tracking and try to fix relative tracking
- Added more debug drawing and options to toggle them off or on.
- Try to fix relative object tracking by verifying the calculations and the operations used to get the inverse transformations.
FTransform.GetRelativeTransform() = To - From where To is the FTransform on the left and From is the FTransform passed to the function.
FTransform.GetRelativeReverse = To - From where From the FTransform on the left and To is the FTransform passed to the function.

NOTICE: seems to work but not sure if it works when the SteamVR transforms are very different from the calibration transforms as during the day the transforms were very different but at the end when testing the SteamVR transforms were very similar to the original calibration transforms all of a sudden again.